use tap, fix ci
