+++ package.json
@@ -6,8 +6,10 @@
   "main": "index.js",
   "types": "types/index.d.ts",
   "scripts": {
-    "test": "tsd && nyc node ./test/index.js",
-    "lint": "standard index.js test/* benchmarks/*"
+    "lint": "standard index.js test/* benchmarks/*",
+    "test": "npm run test:unit && npm run test:typescript",
+    "test:unit": "tap",
+    "test:typescript": "tsd"
   },
   "standard": {
     "ignore": [
@@ -22,18 +24,12 @@
     "http",
     "header"
   ],
-  "files": [
-    "README.md",
-    "LICENSE",
-    "index.js",
-    "types/index.d.ts"
-  ],
   "author": "Aras Abbasi",
   "license": "MIT",
   "devDependencies": {
     "benchmark": "2.1.4",
-    "nyc": "15.1.0",
     "standard": "17.0.0",
+    "tap": "^16.3.0",
     "tsd": "^0.22.0"
   },
   "engines": {
