--- test/index.js
@@ -1,2 +1,0 @@
-require('./negotiate')
-require('./Negotiator')
