+++ test/testCases.js
@@ -1,3 +1,4 @@
+'use strict'
 
 module.exports = [
   ['gzip, , identity', ['deflate', 'gzip'], 'gzip'],
